Move KeychainChunk class from gui to libsync
