+++ src/libsync/creds/keychainchunk.cpp
